Set the  parameter for video streaming to false by default to stop the script at the end of the script + move StreamerTranslation to a separate entity.
