+++ tests/kernel-4.14/test-system-config.xml
@@ -73,8 +73,8 @@ equal to its class name.
 		<rangeSensor commonModule="hcsr04" minValue="0" maxValue="100" filter="median3" />
 		<digitalSensor />
 		<fifo />
-                <iioDevice />
-                <videoDevice />
+        	<iioDevice />
+        	<videoDevice />
 		<lineSensor toleranceFactor="1.0" />
 		<objectSensor toleranceFactor="1.0" />
 		<colorSensor m="3" n="3" />
