+++ trikControl/configs/kernel-4.14/model-config.xml
@@ -110,16 +110,16 @@ Note that device, even if listed here, may fail by itself, then it also will not
 
 	<!-- Video sensor ports. -->
 	<video1>
-	       <videoDevice />
+	    <videoDevice />
 	</video1>
 
-<video2>
-<videoDevice />
+	<video2>
+		<videoDevice />
 	</video2>
 
-<usb-camera>
-<videoDevice />
-</usb-camera>
+	<usb-camera>
+		<videoDevice />
+	</usb-camera>
 	<!-- On-board sensors. -->
 	<!-- If model is not using those, they can be turned off to save system resources, by deleting them or
 		 commenting them out. -->
