+++ trikControl/configs/usb/system-config.xml
@@ -388,13 +388,6 @@ equal to its class name.
 		<colorSensor3x3 class="colorSensor" m="3" n="3" />
 	</deviceTypes>
 
-<deviceGroups>
-<dspSensor>
-			<lineSensor>
-			<objectSensor>
-			<colorSensor>
-		</dspSensor>
-	</deviceGroups>
 	<!-- Format for playSound command, used to play .wav files. %1 designates file name to be played. -->
 	<playWavFile command="aplay --quiet &quot;%1&quot; &amp;" />
