+++ trikControl/configs/v6/model-config.xml
@@ -106,13 +106,13 @@ Note that device, even if listed here, may fail by itself, then it also will not
 	</M4>
 
 	<!-- Video sensor ports. -->
+	<video0>
+		<lineSensor />
+	</video0>
+	
 	<video1>
 		<lineSensor />
 	</video1>
-
-<video2>
-		<lineSensor />
-	</video2>
 
 	<!-- On-board sensors. -->
 	<!-- If model is not using those, they can be turned off to save system resources, by deleting them or
